+++ pkg/repository/queries/diceRoll/cleanOldDiceRolls.sql
@@ -1,6 +1,9 @@
 -- name: CleanOldDiceRolls :exec
 DELETE FROM dice_rolls dr
-WHERE dr.character_id = sqlc.arg(character_id)
+USING characters c
+WHERE c.id = dr.character_id
+  AND c.user_id = sqlc.arg(user_id)
+  AND dr.character_id = sqlc.arg(character_id)
   AND dr.id NOT IN (
       SELECT id FROM dice_rolls
       WHERE character_id = sqlc.arg(character_id)
